examples/agents/python/microsoft-agent-framework/requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.1
agent-framework
Jinja2==3.1.6
